backend/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
defusedxml==0.7.1
dj-database-url==3.1.2
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
idna==3.11
joblib==1.5.3
numpy==2.4.2
oauthlib==3.3.1
pandas==3.0.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
threadpoolctl==3.6.0
urllib3==2.6.3
django-cors-headers
razorpay
django-environ
gunicorn
whitenoise
